+++ Front/src/Components/ExperimentInfo.js
@@ -6,31 +6,12 @@ import { Button } from "react-bootstrap";
 import { Redirect } from "react-router-dom";
 import withReactContent from "sweetalert2-react-content";
 import Swal from "sweetalert2";
-
+import API from "../Api/Api";
 // import Card from "../App/components/MainCard";
 import Aux from "../hoc/_Aux";
 import * as actionTypes from "../store/actions";
 
 import "../styles/homePageStyle.css";
-//import '../assets/scss/themes/bootstrap-overlay/_card.scss'
-
-async function deleteData(url = "", data = {}) {
-  // Default options are marked with *
-  const response = await fetch(url, {
-    method: "DELETE", // *GET, POST, PUT, DELETE, etc.
-    mode: "cors", // no-cors, *cors, same-origin
-    cache: "no-cache", // *default, no-cache, reload, force-cache, only-if-cached
-    credentials: "same-origin", // include, *same-origin, omit
-    headers: {
-      "Content-Type": "application/json",
-      // 'Content-Type': 'application/x-www-form-urlencoded',
-    },
-    redirect: "follow", // manual, *follow, error
-    referrerPolicy: "no-referrer", // no-referrer, *no-referrer-when-downgrade, origin, origin-when-cross-origin, same-origin, strict-origin, strict-origin-when-cross-origin, unsafe-url
-    body: JSON.stringify(data), // body data type must match "Content-Type" header
-  });
-  return response.json(); // parses JSON response into native JavaScript objects
-}
 
 class ExperimentInfo extends Component {
   constructor(props) {
@@ -67,9 +48,8 @@ class ExperimentInfo extends Component {
           questionnaire_id: this.props.chosen.questionnaire_id, //
         };
 
-        deleteData(
-          "http://127.0.0.1:8000/questionnaire-preview-data/" +
-            this.props.chosen.questionnaire_id,
+        API.deleteRequest(
+          "questionnaire-preview-data/" + this.props.chosen.questionnaire_id,
           response
         ).then((data) => {
           console.log(data); // JSON data parsed by `data.json()` call
