+++ Front/src/Components/QuestionsInput.js
@@ -4,44 +4,8 @@ import { connect } from "react-redux";
 import * as actionTypes from "../store/actions";
 // import Card from "../App/components/MainCard";
 import Aux from "../hoc/_Aux";
-// import Breadcrumb from "../App/components/Breadcrumb";
 import { MDBIcon } from "mdbreact";
-
-async function putData(url = "", data = {}) {
-  // Default options are marked with *
-  const response = await fetch(url, {
-    method: "PUT", // *GET, POST, PUT, DELETE, etc.
-    mode: "cors", // no-cors, *cors, same-origin
-    cache: "no-cache", // *default, no-cache, reload, force-cache, only-if-cached
-    credentials: "same-origin", // include, *same-origin, omit
-    headers: {
-      "Content-Type": "application/json",
-      // 'Content-Type': 'application/x-www-form-urlencoded',
-    },
-    redirect: "follow", // manual, *follow, error
-    referrerPolicy: "no-referrer", // no-referrer, *no-referrer-when-downgrade, origin, origin-when-cross-origin, same-origin, strict-origin, strict-origin-when-cross-origin, unsafe-url
-    body: JSON.stringify(data), // body data type must match "Content-Type" header
-  });
-  return response.json(); // parses JSON response into native JavaScript objects
-}
-
-async function deleteData(url = "", data = {}) {
-  // Default options are marked with *
-  const response = await fetch(url, {
-    method: "DELETE", // *GET, POST, PUT, DELETE, etc.
-    mode: "cors", // no-cors, *cors, same-origin
-    cache: "no-cache", // *default, no-cache, reload, force-cache, only-if-cached
-    credentials: "same-origin", // include, *same-origin, omit
-    headers: {
-      "Content-Type": "application/json",
-      // 'Content-Type': 'application/x-www-form-urlencoded',
-    },
-    redirect: "follow", // manual, *follow, error
-    referrerPolicy: "no-referrer", // no-referrer, *no-referrer-when-downgrade, origin, origin-when-cross-origin, same-origin, strict-origin, strict-origin-when-cross-origin, unsafe-url
-    body: JSON.stringify(data), // body data type must match "Content-Type" header
-  });
-  return response.json(); // parses JSON response into native JavaScript objects
-}
+import API from "../Api/Api";
 
 class FormsElements extends React.Component {
   constructor(props) {
@@ -83,8 +47,8 @@ class FormsElements extends React.Component {
       questionnaire_id: this.state.id, //
     };
 
-    putData(
-      "http://127.0.0.1:8000/questionnaire-preview-data/" + this.state.id,
+    API.putRequest(
+      "questionnaire-preview-data/" + this.state.id,
       response
     ).then((data) => {
       // console.log(data); // JSON data parsed by `data.json()` call
@@ -106,8 +70,8 @@ class FormsElements extends React.Component {
       task_id: this.state.taskId, //
     };
 
-    deleteData(
-      "http://127.0.0.1:8000/questionnaire-preview-data/" + this.state.id,
+    API.deleteRequest(
+      "questionnaire-preview-data/" + this.state.id,
       response
     ).then((data) => {
       console.log(data); // JSON data parsed by `data.json()` call
