+++ Front/src/Pages/ExperimentPage.js
@@ -4,24 +4,7 @@ import ComponentsTable from "../Components/Tabels/ComponentsTable";
 import Aux from "../hoc/_Aux";
 import * as actionTypes from "../store/actions";
 import "bootstrap/dist/css/bootstrap.min.css";
-
-async function postData(url = "", data = {}) {
-  // Default options are marked with *
-  const response = await fetch(url, {
-    method: "POST", // *GET, POST, PUT, DELETE, etc.
-    mode: "cors", // no-cors, *cors, same-origin
-    cache: "no-cache", // *default, no-cache, reload, force-cache, only-if-cached
-    credentials: "same-origin", // include, *same-origin, omit
-    headers: {
-      "Content-Type": "application/json",
-      // 'Content-Type': 'application/x-www-form-urlencoded',
-    },
-    redirect: "follow", // manual, *follow, error
-    referrerPolicy: "no-referrer", // no-referrer, *no-referrer-when-downgrade, origin, origin-when-cross-origin, same-origin, strict-origin, strict-origin-when-cross-origin, unsafe-url
-    body: JSON.stringify(data), // body data type must match "Content-Type" header
-  });
-  return response.json(); // parses JSON response into native JavaScript objects
-}
+import API from "../Api/Api";
 
 class ExperimentPage extends Component {
   constructor() {
@@ -59,12 +42,10 @@ class ExperimentPage extends Component {
       questionnaire_type_id: "1", //
     };
 
-    postData("http://127.0.0.1:8000/questionnaire-preview-data", response).then(
-      (data) => {
-        console.log(data); // JSON data parsed by `data.json()` call
-        this.setState({ expId: data.questionnaire_id });
-      }
-    );
+    API.postRequest("questionnaire-preview-data", response).then((data) => {
+      console.log(data); // JSON data parsed by `data.json()` call
+      this.setState({ expId: data.questionnaire_id });
+    });
   }
 
   render() {
